feat: align slack agent engagement dispatch
